# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Qalox Marketer Portal API
  description: |
    Complete REST API for the Marketer Portal, providing authentication, wallet management,
    school management, commission tracking, and analytics endpoints.

    **Authentication:** All endpoints require `x-service-key` header for service-to-service authentication.
  version: 1.0.0
  contact:
    name: Qalox Backend Team
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/public
    description: Development Server
  - url: https://api.qalox.com/api/public
    description: Production Server
...
components:
  securitySchemes:
    serviceKey:
      type: apiKey
      in: header
      name: x-service-key
      description: Service-to-service authentication key
  parameters:
    marketerId:
      name: marketerId
      in: query
      required: true
      schema:
        type: integer
      description: ID of the marketer (Admin user with role=marketer)
    page:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
      description: Page number for pagination
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Items per page
    startDate:
      name: startDate
      in: query
      schema:
        type: string
        format: date-time
      description: Filter from this date (ISO 8601 format)
    endDate:
      name: endDate
      in: query
      schema:
        type: string
        format: date-time
      description: Filter until this date (ISO 8601 format)
  schemas:
    Error:
      type: object
      required:
        - success
        - message
        - code
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Resource not found"
        code:
          type: string
          example: "NOT_FOUND"
...
security:
  - serviceKey: []
paths:
...
  /marketers:
    get:
      tags:
        - Super Admin - Platform Management
      summary: List all marketers (service-to-service)
      security:
        - serviceKey: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    description: Nested shape consumed by qaloxApiClient.listMarketers
                    properties:
                      data:
                        type: array
                        items: { type: object }
                      meta:
                        type: object
                        properties:
                          page: { type: integer }
                          limit: { type: integer }
                          total: { type: integer }
                          totalPages: { type: integer }
  /marketers/{id}/status:
    patch:
      tags:
        - Super Admin - Platform Management
      summary: Suspend/activate a marketer (alias)
      description: |
        Alias of `PATCH /marketers/{id}/suspend` — the Super Admin Portal's
        `qaloxApiClient.toggleMarketerStatus` calls this path specifically.
        Both routes point to the same handler.
      security:
        - serviceKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                suspend: { type: boolean }
                reason: { type: string, nullable: true }
      responses:
        '200':
          description: Success
        '404':
          description: Marketer not found
  /schools/{id}/sms-quota:
    patch:
      tags:
        - Super Admin - Platform Management
      summary: Set a school's per-term SMS broadcast quota
      description: |
        Super Admin control over each school's Bulk SMS Broadcast allowance
        (Phase D). Applied directly to the School record for now — once
        schools are formally linked to a BillingPlan, this should be derived
        from the plan's quota instead of set ad-hoc per school.
      security:
        - serviceKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quotaPerTerm]
              properties:
                quotaPerTerm: { type: integer, minimum: 0, example: 300 }
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      id: { type: integer }
                      name: { type: string }
                      smsQuotaPerTerm: { type: integer }
                      smsUsedThisTerm: { type: integer }
        '400':
          description: Invalid quotaPerTerm
        '404':
          description: School not found
...
  /auth/2fa/verify:
    post:
      tags:
        - Phase 1 - Core
      summary: Verify 2FA code during login
      description: Validates 2FA OTP code and returns authenticated user token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tempToken
                - code
              properties:
                tempToken:
                  type: string
                  description: Temporary JWT token from login
                code:
                  type: string
                  pattern: '^\d{6}$'
                  description: 6-digit OTP code
      responses:
        '200':
          description: 2FA verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      email:
                        type: string
                      name:
                        type: string
                      token:
                        type: string
        '400':
          description: Invalid or missing fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /marketers/{marketerId}/wallet:
    post:
      tags:
        - Phase 1 - Core
      summary: Process wallet operation (withdraw/credit)
      description: Perform withdraw or credit operation on marketer wallet
      parameters:
        - name: marketerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - amount
                - operation
              properties:
                amount:
                  type: number
                  format: float
                  minimum: 0
                operation:
                  type: string
                  enum: [withdraw, credit]
      responses:
        '200':
          description: Operation successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      marketerId:
                        type: integer
                      balance:
                        type: number
                      totalWithdrawn:
                        type: number
                      lastPayoutDate:
                        type: string
                        format: date-time
        '400':
          description: Invalid operation or insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Marketer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /school-tokens/{id}/revoke:
    patch:
      tags:
        - Phase 1 - Core
      summary: Revoke a school token
      description: Deactivate/revoke an active token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Token revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      status:
                        type: string
        '404':
          description: Token not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /commissions:
    get:
      tags:
        - Phase 1 - Core
      summary: Get marketer commissions
      description: Retrieve paginated list of commissions with optional date filtering. marketerId is extracted from JWT token.
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/startDate'
        - $ref: '#/components/parameters/endDate'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      commissions:
                        type: array
                        items:
                          $ref: '#/components/schemas/Commission'
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      pages:
                        type: integer
...
  /commissions/monthly-chart:
    get:
      tags:
        - Phase 3 - Advanced
      summary: Get monthly commission chart data
      description: Retrieve commission data for last N months for charting. marketerId is extracted from JWT token.
      parameters:
        - name: months
          in: query
          schema:
            type: integer
            default: 12
            minimum: 1
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        month:
                          type: string
                        amount:
                          type: number
  /notifications/{id}/read:
    put:
      tags:
        - Phase 1 - Core
      summary: Mark notification as read
      description: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
...
  /users/profile:
    put:
      tags:
        - Phase 2 - Profile & Settings
      summary: Update marketer profile
      description: Update profile information (name, phone, address, city, state)
      parameters:
        - $ref: '#/components/parameters/marketerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                address:
                  type: string
                city:
                  type: string
                state:
                  type: string
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Marketer'
  /users/password:
    put:
      tags:
        - Phase 2 - Profile & Settings
      summary: Change password
      description: Change marketer password with current password verification
      parameters:
        - $ref: '#/components/parameters/marketerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  format: password
                newPassword:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Invalid current password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/avatar:
    post:
      tags:
        - Phase 2 - Profile & Settings
      summary: Upload avatar
      description: Upload profile avatar image
      parameters:
        - $ref: '#/components/parameters/marketerId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                avatar:
                  type: string
                  format: binary
      responses:
        '200':
          description: Avatar uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      avatarUrl:
                        type: string
...
  /settings/verify-account:
    get:
      tags:
        - Phase 2 - Profile & Settings
      summary: Verify bank account
      description: Verify bank account number and name using bank verification service
      parameters:
        - name: accountNumber
          in: query
          required: true
          schema:
            type: string
        - name: bankCode
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Account verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accountNumber:
                        type: string
                      accountName:
                        type: string
                      bankCode:
                        type: string
                      verified:
                        type: boolean
...